fixing deployment (WIP)
