+++ src/components/Header.jsx
@@ -1,12 +1,17 @@
 import React, {useState} from 'react';
 import LoginModal from './LoginModal';
-import { Header, Segment } from 'semantic-ui-react';
+import { Container, Header, Segment, Form, Button } from 'semantic-ui-react';
 import BecomeSubscriber from './BecomeSubscriber';
 import { Elements } from 'react-stripe-elements';
 
 const HeaderMainPage = (props) => {
 	const [authenticated, setAuthenticated] = useState(false)
   const [subscribed, setSubscribed] = useState(false)
+
+	const onSubmit = (event) => {
+    event.preventDefault();
+    props.setQuery(event.target.search.value);
+  };
 
   return (
     <div data-cy='header'>
@@ -24,6 +29,19 @@ const HeaderMainPage = (props) => {
             <BecomeSubscriber setMessage={props.setMessage} setSubscribed={setSubscribed}/>
           </Elements>
         )}
+				<Container position='centered'>
+        {subscribed && <Form onSubmit={(event) => onSubmit(event)}>
+          <Form.Input
+            type='text'
+            name='search'
+            placeholder='Search Movie...'
+            data-cy='search-input'
+          />
+          <Button type='submit' data-cy='search-btn'>
+            Search
+          </Button>
+        </Form>}
+      </Container>
       </Segment>
     </div>
   );
